Added introduce functionality.
Note that `test_intros.py` reveals some unfriendly
aspects of client certificate validation.
We can't trust a client certificate directly,
since it's not self-signed.  We also can't add the
client certificate we've tried to our trust store,
since this isn't the exact one the client will send.

So, we have to resort to one of the following:

1. Have the client use this specific introduction letter
   (certificate chain) when talking to this specific server.

2. Add the client's signing certificate to known_clients,
   which means we'll trust all certificates the client signs.

3. (untested) always have the client present a self-signed
   certificate.  This is poor design since then the same
   key would be used for authentication and signing,
   and since it doesn't allow any chains of trust.

It doesn't seem possible to have the server directly
trust any given end-entity certificate.  The client
has to present the server a certificate signed by
the "right" party.